refactor (addin): Breaking up the UIBindings class.
Specialized code modules for Speckle handling, Conversions and Geometry.
