+++ topics/addin/Rimshot.2022/Properties/AssemblyInfo.cs
